Add some entries to .gitignore
To avoid adding things that are often lying around...

The .clj and .glj are to ignore top level scripts for testing stuff.

The GNUmakefile and .cache are for alternate Makefile stuff like I do in
https://github.com/glojurelang/glojure/pull/75

If that doesn't get merged I can just symlink to it since GNUmakefile
has precedence over Makefile (for GNU make).