Makefile use bash for all commands
Setting SHELL to bash allows us to make use of a more dependable
environment. It also allows us to use bash syntax which in many cases is
cleaner because you don't need to be concerned with supporting any
shell.

`bash` is preferable to `/bin/bash` since it will pick up possibly more
preferable bash version from the PATH. Like on macOS, /bin/bash is stuck
at 3.2, but its easy to `brew install bash` to get the latest version in
/usr/local/bin/bash.